Put CondensedUserView into their own class (IOS-141)
