+++ Mastodon/Scene/Search/SearchDetail/SearchHistory/SearchHistorySection.swift
@@ -31,7 +31,7 @@ extension SearchHistorySection {
         let userCellRegister = UICollectionView.CellRegistration<SearchHistoryUserCollectionViewCell, ManagedObjectRecord<MastodonUser>> { cell, indexPath, item in
             context.managedObjectContext.performAndWait {
                 guard let user = item.object(in: context.managedObjectContext) else { return }
-                cell.configure(with: user)
+                cell.condensedUserView.configure(with: user)
             }
         }
